adding more functions to manage features, types and sizes; creating sets of features from rosbag demonstrations

+++ include/grid/test_features.h
@@ -16,14 +16,14 @@ namespace grid {
      * A feature query gets the set of all featutes for different points in time, normalizes them, and returns.
      */
     std::vector<Pose> getPose(const std::string &name,
-                              unsigned long int mintime = 0,
-                              unsigned long int maxtime = 0);
+                              double mintime = 0,
+                              double maxtime = 0);
     /* getFeatureValues
      * Returns a list of features converted into a format we can use.
      */
     std::vector<FeatureVector> getFeatureValues(const std::string &name,
-                                                unsigned long int mintime = 0,
-                                                unsigned long int maxtime = 1);
+                                                double mintime = 0,
+                                                double maxtime = 1);
 
     /* getFeaturesForTrajectory
      * Get information for a single feature over the whole trajectory given in traj.
